Re-group integration cards
